Created FormatterComponent & formatters can be disabled
It allows disabling chat processing as a whole or formatting
Formatters can be disabled one by one
#85
