+++ src/main/java/buttondevteam/chat/ChatUtils.java
@@ -1,11 +1,19 @@
 package buttondevteam.chat;
 
+import buttondevteam.lib.TBMCChatEvent;
+import buttondevteam.lib.ThorpeUtils;
+import lombok.var;
 import org.bukkit.Bukkit;
+import org.bukkit.entity.Player;
 
 import java.util.Optional;
+import java.util.function.Function;
 
 public final class ChatUtils {
-	private ChatUtils() {}
+	public static final String MCORIGIN = "Minecraft"; //Shouldn't change, like ever - TBMCPlayer.getFolderForType(TBMCPlayer.class) capitalized
+
+	private ChatUtils() {
+	}
 
 	/**
 	 * Dispatch a console command.
@@ -32,5 +40,21 @@ public final class ChatUtils {
 		int a = str.indexOf(start) + start.length();
 		int b = str.indexOf(end, a);
 		return a != -1 && b != -1 ? Optional.of(str.substring(a, b)) : Optional.empty();
+	}
+
+	/**
+	 * Sends a regular (non-Markdown) chat message. Used as a fallback if the chat processing fails.
+	 *
+	 * @param e        The chat event
+	 * @param modifier A function that alters the message to be displayed to the player
+	 */
+	public static void sendChatMessage(TBMCChatEvent e, Function<String, String> modifier) {
+		var str = "[" + e.getChannel().DisplayName().get() + "] <"
+			+ ThorpeUtils.getDisplayName(e.getSender()) + "> " + e.getMessage();
+		str = modifier.apply(str);
+		for (Player p : Bukkit.getOnlinePlayers())
+			if (e.shouldSendTo(p))
+				p.sendMessage(str);
+		Bukkit.getConsoleSender().sendMessage(str);
 	}
 }
